Bump deps

+++ backend/src/api/cacheable.rs
@@ -1,7 +1,6 @@
 use std::fmt::Display;
 
 use anyhow::anyhow;
-use async_trait::async_trait;
 use chrono::{DateTime, Duration, Utc};
 use redis::{AsyncCommands, SetOptions, ToRedisArgs};
 use serde::de::DeserializeOwned;
@@ -44,7 +43,6 @@ impl ToRedisArgs for CacheKey {
     }
 }
 
-#[async_trait]
 pub trait Cacheable: DeserializeOwned + serde::Serialize {
     fn cache_key() -> CacheKey;
     fn expires_at() -> DateTime<Utc>;
@@ -123,7 +121,7 @@ pub trait Cacheable: DeserializeOwned + serde::Serialize {
         // let unwrapped_data = *data.lock().await;
         let json_data = serde_json::to_string(&*data.lock().await)?;
 
-        let unix_timestamp_expiry = usize::try_from(Self::expires_at().timestamp())?;
+        let unix_timestamp_expiry = u64::try_from(Self::expires_at().timestamp())?;
         let options =
             SetOptions::default().with_expiration(redis::SetExpiry::EXAT(unix_timestamp_expiry));
